requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
distro==1.9.0
Django==5.2.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
lxml==6.0.2
openai==2.6.1
pydantic==2.12.3
pydantic_core==2.41.4
python-docx==1.2.0
python-dotenv==1.2.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
